+++ middleware/src/modules/livekit/livekit.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { AvatarController } from './avatar.controller';
+import { LiveKitService } from './livekit.service';
+
+@Module({
+  controllers: [AvatarController],
+  providers: [LiveKitService],
+  exports: [LiveKitService],
+})
+export class LiveKitModule {}